SIM-2893: Remove a hard depenency on langchain_openai.
* Explicitly tell mypy to ignore the langchain_openai dependency.
* Add Warning when using openAI by default.
* Update README Qickstart to add 'poetry install --with openai'
  so that the existing behaviour is preserved.
* Move demo simulation code to just be
  > fable_saga.demos.space_colony (removed the last simulation part)
* Fix test not using fake_openai as a parameter.
