+++ fable_saga/demos/space_colony/__main__.py
@@ -1,0 +1,4 @@
+import asyncio
+from . import main
+
+asyncio.run(main())
