+++ mypy.ini
@@ -6,3 +6,6 @@ mypy_path = ./stubs
 check_untyped_defs = True
 # This is needed to allow pycharm to find the location of the error automatically.
 show_column_numbers = True
+
+[mypy-langchain_openai]
+ignore_missing_imports = True
