+++ stubs/engineio/async_drivers/gevent.pyi
@@ -1,14 +1,11 @@
-import gevent
 from _typeshed import Incomplete
 from engineio.async_drivers._websocket_wsgi import SimpleWebSocketWSGI as SimpleWebSocketWSGI
 
-SimpleWebSocketWSGI: Incomplete
+gevent: Incomplete
 
 class Thread(gevent.Greenlet):
     def __init__(self, target, args=[], kwargs={}) -> None: ...
 
-class WebSocketWSGI(SimpleWebSocketWSGI):
-    def __init__(self, handler, server) -> None: ...
 
 class WebSocketWSGI:
     app: Incomplete
